refactor(client): split chat_provider into parts
1295 LOC monolith split into a chat/ folder. New layout:
- chat_provider.dart (6 LOC) — re-export shim so existing imports
  keep working unchanged.
- chat/chat_provider.dart (403) — public ChatNotifier + facade.
- chat/chat_state.dart (300) — ChatState immutable class.
- chat/chat_history.dart (389) — load/append/paginate cache.
- chat/chat_edits.dart (126) — delete-for-me, edits.
- chat/chat_reactions.dart (61) — reaction state mutations.
- chat/chat_recovery.dart (92) — decrypt-failure recovery.

Behaviour-preserving. flutter analyze clean. All 1921 client tests
pass — including the 48 chat_provider_test.dart unit tests.
